Change power_trace's index for exceptions
Because exceptions' the power_trace will be called from the TracePoint
block instead of the actual method that causes the error, power_trace
will have an extra but confusing entry than the normal backtrace. So we
need to advance the index value by 1 if the power_trace is called for an
exception.