Add adjust_rspec_error global config
Because RSpec indent error's backtrace automatically, it'll make the
error messages look like

```
   # /Users/st0012/projects/ticketsolve/app/modules/merging/services/customer_merging_service.rb:56:in `block in destroy_duplicated_null_customer!'
(Locals) <- these are not indented
  model: Member
  customer_id: 1029
  duplicated_null_customer: #<NullCustomer:0x00007ff0003e4570>
```

To solve this problem, we'll need to tell power_trace to increase
additional information sections' indentation, before we convert entries
into string format. And currently the best place to do it is inside
RSpec's ExceptionPresenter class.

So if the `PowerTrace.adjust_rspec_error` is set to `true`,
ExceptionPresenter will use `exception.power_trace.to_backtrace` instead
of `exception.backtrace` to extract the backtrace to display.

Another thing to keep in mind is that if `PowerTrace.replace_backtrace`
is enabled, this behavior will be enforced as well to keep the error
message looks consistent.