Avoid restoring exceptions' power_trace
In some usecases, like ActiveRecord's `within_new_transaction` method,
exceptions could be rescued and re-raised.

And only the initial error raising havve the accurate bindings that can provide useful
information. If we capture the power_trace when the exception is
re-raised, the trace won't be accurate.

Because of this, we should only collect and store power trace when an
exception is firstly raised.